pyproject.toml
[project]
name = "easyorder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.135.2",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-logging>=0.58b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.58b0",
    "opentelemetry-sdk>=1.37.0",
    "bcrypt>=4.0.0",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.8",
    "semgrep>=1.156.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["F", "E", "I", "N", "UP", "B", "SIM", "RUF", "S", "DTZ", "T10", "T20", "ERA", "A", "RET", "C4", "PLE", "C90"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # 测试中允许 assert
"app/api/**" = ["B008"]  # FastAPI Depends() 需要在参数默认值中调用

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
